Rewrote the Display class to make it more concise. Trim down the Scene class even further by removing the display related code. Also, updated the other code snippets associated with the display.

+++ Pathfinder/source/Game/Algorithms/Algorithm.h
@@ -21,7 +21,7 @@ public:
 
 public:
 	virtual bool Execute(AlgorithmType algorithmType) override;
-	virtual bool PlayVisualization(float speed, float deltaTime) override;
+	virtual bool PlayVisualization(int speed, float deltaTime) override;
 	virtual void Stop() override;
 
 	virtual bool IsPathFound() override;
